requirements.txt
disnake
configparser
sqlalchemy
psycopg2
python-dotenv